Source code for 8A Assignment
I'm having problems with my map showing up in the application when I open it in Chrome. I've used the developer tools in the browser for de-bugging, and under "Watch" it tells me that my operationalLayer url is unavailable, which isn't surprising because the legend and gauge diijts are appearing but not the map. I've examined the Hurricanes layer url many times, opened it in different windows, even tried using the pjson version of the layer but with no luck. Is this an issue on the browser/development environment side, or what? I've tried zooming too.